TelegramBot/old/requirements.txt
python-telegram-bot==12.7
torch == 1.8.0
transformers == 4.4.2
